Can read customers from file, and try to release them on time, but it doesn't seem to be adding to the queues

+++ p2/Queue.h
@@ -8,8 +8,8 @@
 typedef struct customer{
 	int id;
 	int business; 
-	int arrivalTime;
-	int serviceTime; 
+	float arrivalTime;
+	float serviceTime; 
     struct customer *next; 
 }customer;
